Make cache control follow our standard built-in plugin pattern
That means:
- Remove standalone `apollo-cache-control` package. Move the code into
  `apollo-server-core/src/plugin` and some of the types into
  `apollo-server-types`. Make its extensions to `GraphQLRequestContext`
  just be part of the definition of that type.
- Drop top-level `cacheControl` constructor option; instead, use our
  now-standard pattern of "you can put the extension in `plugins`
  yourself, you can put the 'Disabled' version in `plugins`, or
  otherwise AS will add the plugin with default configuration".

There's no "semantic" change in this PR: it just changes how you write
`cacheControl: false` and `cacheControl: options` to
    plugins: [ApolloServerPluginCacheControlDisabled()]
and
    plugins: [ApolloServerPluginCacheControl(options)]
respectively.

When I wrote this commit, I started to get concerned about the ordering
of the cache control plugin (which used to be always first... well,
other than an implicit usage reporting plugin, which adds itself with
this.plugin.unshift) with respect to other plugins like usage reporting
and the response cache plugin. What if it wrote to
requestContext.overallCachePolicy too late? In practice, the ordering
doesn't appear to matter, but see a later commit in this PR for a change
that makes this issue moot by continuously updating overallCachePolicy.
