+++ package.json
@@ -27,7 +27,6 @@
   },
   "dependencies": {
     "@apollographql/apollo-tools": "0.5.0",
-    "apollo-cache-control": "file:packages/apollo-cache-control",
     "apollo-datasource": "file:packages/apollo-datasource",
     "apollo-datasource-rest": "file:packages/apollo-datasource-rest",
     "apollo-reporting-protobuf": "file:packages/apollo-reporting-protobuf",
