--- packages/apollo-cache-control/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "apollo-cache-control",
-  "version": "0.300.0-alpha.4",
-  "description": "A GraphQL extension for cache control",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "license": "MIT",
-  "repository": "apollographql/apollo-cache-control-js",
-  "author": "Apollo <opensource@apollographql.com>",
-  "engines": {
-    "node": ">=12.0"
-  },
-  "dependencies": {
-    "apollo-server-env": "file:../apollo-server-env",
-    "apollo-server-plugin-base": "file:../apollo-server-plugin-base"
-  },
-  "peerDependencies": {
-    "graphql": "^15.3.0"
-  }
-}
