--- packages/apollo-cache-control/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "extends": "../../tsconfig.base",
-  "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "./dist"
-  },
-  "include": ["src/**/*"],
-  "exclude": ["**/__tests__"],
-  "references": [
-    { "path": "../apollo-server-plugin-base" },
-  ]
-}
