+++ packages/apollo-server-core/package.json
@@ -31,7 +31,6 @@
     "@graphql-tools/schema": "^7.1.5",
     "@graphql-tools/utils": "^7.9.0",
     "@josephg/resolvable": "^1.0.0",
-    "apollo-cache-control": "file:../apollo-cache-control",
     "apollo-datasource": "file:../apollo-datasource",
     "apollo-graphql": "^0.9.0",
     "apollo-reporting-protobuf": "file:../apollo-reporting-protobuf",
