+++ packages/apollo-server-core/src/__tests__/runQuery.test.ts
@@ -64,7 +64,6 @@ function runQuery(
 interface QueryOptions
   extends Pick<
     GraphQLOptions<GraphQLContext<any>>,
-    | 'cacheControl'
     | 'context'
     | 'debug'
     | 'documentStore'
