+++ packages/apollo-server-core/src/plugin/cacheControl/__tests__/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../../../../../tsconfig.test.base",
+  "include": ["**/*"],
+  "references": [
+    { "path": "../../../../" },
+  ]
+}
