+++ packages/apollo-server-core/tsconfig.json
@@ -7,7 +7,6 @@
   "include": ["src/**/*"],
   "exclude": ["**/__tests__"],
   "references": [
-    { "path": "../apollo-cache-control" },
     { "path": "../apollo-datasource" },
     { "path": "../apollo-server-caching" },
     { "path": "../apollo-server-errors" },
