+++ packages/apollo-server-fastify/src/__tests__/ApolloServer.test.ts
@@ -5,7 +5,7 @@ import http from 'http';
 
 import request from 'request';
 
-import { gql, AuthenticationError, Config } from 'apollo-server-core';
+import { gql, AuthenticationError, Config, ApolloServerPluginCacheControlDisabled } from 'apollo-server-core';
 import { ApolloServer, ServerRegistration } from '../ApolloServer';
 
 import {
@@ -687,7 +687,7 @@ describe('apollo-server-fastify', () => {
         const { url: uri } = await createServer({
           typeDefs,
           resolvers,
-          cacheControl: false,
+          plugins: [ApolloServerPluginCacheControlDisabled()],
         });
 
         const apolloFetch = createApolloFetch({ uri }).useAfter(
