+++ packages/apollo-server-plugin-response-cache/package.json
@@ -20,7 +20,6 @@
     "node": ">=12.0"
   },
   "dependencies": {
-    "apollo-cache-control": "file:../apollo-cache-control",
     "apollo-server-caching": "file:../apollo-server-caching",
     "apollo-server-plugin-base": "file:../apollo-server-plugin-base"
   },
