+++ packages/apollo-server-plugin-response-cache/tsconfig.json
@@ -7,7 +7,6 @@
   "include": ["src/**/*"],
   "exclude": ["**/__tests__"],
   "references": [
-    { "path": "../apollo-cache-control" },
     { "path": "../apollo-server-plugin-base" },
     { "path": "../apollo-server-caching" },
     { "path": "../apollo-server-types" },
