+++ tsconfig.test.json
@@ -5,7 +5,6 @@
   "files": [],
   "include": [],
   "references": [
-    { "path": "./packages/apollo-cache-control/src/__tests__/" },
     { "path": "./packages/apollo-datasource-rest/src/__tests__/" },
     { "path": "./packages/apollo-server/src/__tests__/" },
     { "path": "./packages/apollo-server-azure-functions/src/__tests__/" },
@@ -14,6 +13,8 @@
     { "path": "./packages/apollo-server-caching/src/__tests__/" },
     { "path": "./packages/apollo-server-cloud-functions/src/__tests__/" },
     { "path": "./packages/apollo-server-core/src/__tests__/" },
+    { "path": "./packages/apollo-server-core/src/plugin/cacheControl/__tests__/" },
+    { "path": "./packages/apollo-server-core/src/plugin/usageReporting/__tests__/" },
     { "path": "./packages/apollo-server-errors/src/__tests__/" },
     { "path": "./packages/apollo-server-express/src/__tests__/" },
     { "path": "./packages/apollo-server-fastify/src/__tests__/" },
